# docs/swagger.yaml
basePath: /api
definitions:
...
  dtos.CreateCommentRequest:
    properties:
      content:
        example: This is a great task!
        maxLength: 150
        minLength: 3
        type: string
    required:
    - content
    type: object
...
  dtos.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        example: password123
        maxLength: 12
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dtos.RegisterRequest:
    properties:
      email:
        example: user@example.com
        type: string
      fullName:
        example: John Doe
        maxLength: 100
        minLength: 6
        type: string
      gender:
        enum:
        - male
        - female
        example: male
        type: string
      password:
        example: password123
        maxLength: 12
        minLength: 6
        type: string
    required:
    - email
    - fullName
    - gender
    - password
    type: object
...
  errors.BasicError:
    properties:
      error:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  models.ActivityLog:
    properties:
      action:
        type: string
      created_at:
        type: string
      id:
        type: string
      meta_data:
        items:
          type: integer
        type: array
      org_id:
        type: string
      project_id:
        type: string
      task_id:
        type: string
      user_id:
        type: string
    required:
    - action
    type: object
  models.Comment:
    properties:
      content:
        type: string
      created_at:
        type: string
      id:
        type: string
      task_id:
        type: string
      user_id:
        type: string
    required:
    - content
    - task_id
    - user_id
    type: object
...
  models.UserResponse:
    properties:
      created_at:
        type: string
      email:
        example: user@example.com
        type: string
      fullName:
        example: John Doe
        type: string
      gender:
        example: male
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      imageUrl:
        example: avatar.png
        type: string
      last_login:
        type: string
      updated_at:
        type: string
    type: object
  utils.SuccessResponse:
    properties:
      data: {}
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
    type: object
  utils.SuccessResponseWithTotal:
    properties:
      data: {}
      message:
        type: string
      path:
        type: string
      success:
        type: boolean
      timestamp:
        type: string
      total:
        type: integer
    type: object
host: localhost:3000
info:
  contact:
    email: support@taskio.com
    name: API Support
  description: Task management system API with real-time features
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Taskio API
  version: "1.0"
paths:
  /activities:
    get:
      consumes:
      - application/json
      description: Retrieves all activity logs for the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: Activities fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponseWithTotal'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.ActivityLog'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get user activities
      tags:
      - Activities
  /activities/{activityId}:
    delete:
      consumes:
      - application/json
      description: Deletes a specific activity log by ID
      parameters:
      - description: Activity ID (UUID)
        in: path
        name: activityId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Activity deleted successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "400":
          description: Invalid activity ID
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Delete an activity
      tags:
      - Activities
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates a user and returns a JWT token in cookie
      parameters:
      - description: Login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Logged in successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      summary: User login
      tags:
      - Auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates the refresh token and clears the authentication cookie
      produces:
      - application/json
      responses:
        "200":
          description: Logged out successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: User logout
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the authentication token using the refresh token in cookie
      produces:
      - application/json
      responses:
        "200":
          description: Token refreshed successfully
          schema:
            $ref: '#/definitions/utils.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Refresh JWT token
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user account
      parameters:
      - description: User registration details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Registered successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.UserResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      summary: User registration
      tags:
      - Auth
...
  /tasks/{taskId}/comments:
    get:
      consumes:
      - application/json
      description: Retrieves all comments for a specific task
      parameters:
      - description: Task ID (UUID)
        in: path
        name: taskId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Comments fetched successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.Comment'
                  type: array
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Get task comments
      tags:
      - Comments
    post:
      consumes:
      - application/json
      description: Adds a new comment to a specific task
      parameters:
      - description: Task ID (UUID)
        in: path
        name: taskId
        required: true
        type: string
      - description: Comment content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dtos.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Comment created successfully
          schema:
            allOf:
            - $ref: '#/definitions/utils.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/models.Comment'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/errors.BasicError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errors.BasicError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/errors.BasicError'
      security:
      - BearerAuth: []
      summary: Create a comment
      tags:
      - Comments
...
securityDefinitions:
  BearerAuth:
    description: JWT token stored in cookie
    in: cookie
    name: taskio-cookie
    type: apiKey
swagger: "2.0"